+++ CHANGELOG.md
@@ -17,6 +17,9 @@ with respect to its command line interface and HTTP interface
 * Client: sous jenkins cli to use manifest metadata to generate JenkinsPipeline file
 * Client: more accurate error message on 'sous manifest set' when ManifestID in YAML
   does not match that specified by flags and context.
+* Client: 'artifact add' now records docker image refs using the configured default
+  docker registry name. Previously it sometimes recorded local image refs which
+  were not deployable.
 
 ### Changed
 * Client: `SOUS_USE_SOUS_SEVER` env var must now be exactly uppercase `YES`